requirements.txt
aiohttp==3.5.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
feedparser==5.2.1
future==0.17.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
persian==0.4.0
psycopg2-binary==2.8.2
pycparser==2.19
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.4
typing-extensions==3.7.2
urllib3==1.25.3
yarl==1.3.0
